+++ actors/tests/datacap_test.go
@@ -1,0 +1,191 @@
+package actortest
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/filecoin-project/go-state-types/manifest"
+	filTypes "github.com/filecoin-project/lotus/chain/types"
+	"github.com/ipfs/go-cid"
+	"github.com/stretchr/testify/require"
+	actorsV1 "github.com/zondax/fil-parser/actors/v1"
+	actorsV2 "github.com/zondax/fil-parser/actors/v2"
+	"github.com/zondax/fil-parser/parser"
+)
+
+var datacapWithParamsAndReturnTests = []struct {
+	name   string
+	txType string
+}{
+	{
+		name:   "Mint Exported",
+		txType: parser.MethodMintExported,
+	},
+	{
+		name:   "Burn Exported",
+		txType: parser.MethodBurnExported,
+	},
+	{
+		name:   "Balance Exported",
+		txType: parser.MethodBalanceExported,
+	},
+	{
+		name:   "Transfer Exported",
+		txType: parser.MethodTransferExported,
+	},
+	{
+		name:   "Transfer From Exported",
+		txType: parser.MethodTransferFromExported,
+	},
+	{
+		name:   "Destroy Exported",
+		txType: parser.MethodDestroyExported,
+	},
+	{
+		name:   "Increase Allowance Exported",
+		txType: parser.MethodIncreaseAllowanceExported,
+	},
+	{
+		name:   "Decrease Allowance Exported",
+		txType: parser.MethodDecreaseAllowanceExported,
+	},
+	{
+		name:   "Revoke Allowance Exported",
+		txType: parser.MethodRevokeAllowanceExported,
+	},
+	{
+		name:   "Burn From Exported",
+		txType: parser.MethodBurnFromExported,
+	},
+	{
+		name:   "Allowance Exported",
+		txType: parser.MethodAllowanceExported,
+	},
+}
+
+var datacapWithParamsOrReturnTests = []struct {
+	name   string
+	txType string
+	key    string
+}{
+	{
+		name:   "Name Exported",
+		txType: parser.MethodNameExported,
+		key:    parser.ReturnKey,
+	},
+	{
+		name:   "Symbol Exported",
+		txType: parser.MethodSymbolExported,
+		key:    parser.ReturnKey,
+	},
+	{
+		name:   "Total Supply Exported",
+		txType: parser.MethodTotalSupplyExported,
+		key:    parser.ReturnKey,
+	},
+	{
+		name:   "Granularity Exported",
+		txType: parser.MethodGranularityExported,
+		key:    parser.ReturnKey,
+	},
+}
+
+func TestActorParserV1_DatacapWithParamsAndReturn(t *testing.T) {
+	p := getActorParser(actorsV1.NewActorParser).(*actorsV1.ActorParser)
+
+	for _, tt := range datacapWithParamsAndReturnTests {
+		t.Run(tt.name, func(t *testing.T) {
+			rawParams, rawReturn, err := getParamsAndReturn(manifest.DatacapKey, tt.txType)
+			require.NoError(t, err)
+			require.NotNil(t, rawParams)
+			require.NotNil(t, rawReturn)
+
+			got, err := p.ParseDatacap(tt.txType, &parser.LotusMessage{
+				Params: rawParams,
+			}, &parser.LotusMessageReceipt{
+				Return: rawReturn,
+			})
+			require.NoError(t, err)
+			require.NotNil(t, got)
+			require.Contains(t, got, parser.ParamsKey, "Params could no be found in metadata")
+			require.NotNil(t, got[parser.ParamsKey])
+			require.Contains(t, got, parser.ReturnKey, "Return could no be found in metadata")
+			require.NotNil(t, got[parser.ReturnKey])
+		})
+	}
+}
+
+func TestActorParserV1_DatacapWithParamsOrReturn(t *testing.T) {
+	p := getActorParser(actorsV1.NewActorParser).(*actorsV1.ActorParser)
+
+	for _, tt := range datacapWithParamsOrReturnTests {
+		t.Run(tt.name, func(t *testing.T) {
+			rawParams, err := loadFile(manifest.DatacapKey, tt.txType, tt.key)
+			require.NoError(t, err)
+			require.NotNil(t, rawParams)
+
+			got, err := p.ParseDatacap(tt.txType, &parser.LotusMessage{
+				Params: rawParams,
+			}, &parser.LotusMessageReceipt{
+				Return: nil,
+			})
+			require.NoError(t, err)
+			require.NotNil(t, got)
+			require.Contains(t, got, tt.key, fmt.Sprintf("%s could no be found in metadata", tt.key))
+			require.NotNil(t, got[tt.key])
+		})
+	}
+}
+
+func TestActorParserV2_DatacapWithParamsAndReturn(t *testing.T) {
+	p := getActorParser(actorsV2.NewActorParser).(*actorsV2.ActorParser)
+	actor, err := p.GetActor(manifest.DatacapKey)
+	require.NoError(t, err)
+	require.NotNil(t, actor)
+
+	for _, tt := range datacapWithParamsAndReturnTests {
+		t.Run(tt.name, func(t *testing.T) {
+			rawParams, rawReturn, err := getParamsAndReturn(manifest.DatacapKey, tt.txType)
+			require.NoError(t, err)
+			require.NotNil(t, rawParams)
+			require.NotNil(t, rawReturn)
+
+			got, _, err := actor.Parse(network, height, tt.txType, &parser.LotusMessage{
+				Params: rawParams,
+			}, &parser.LotusMessageReceipt{
+				Return: rawReturn,
+			}, cid.Undef, filTypes.EmptyTSK)
+			require.NoError(t, err)
+			require.NotNil(t, got)
+			require.Contains(t, got, parser.ParamsKey, "Params could no be found in metadata")
+			require.NotNil(t, got[parser.ParamsKey])
+			require.Contains(t, got, parser.ReturnKey, "Return could no be found in metadata")
+			require.NotNil(t, got[parser.ReturnKey])
+		})
+	}
+}
+
+func TestActorParserV2_DatacapWithParamsOrReturn(t *testing.T) {
+	p := getActorParser(actorsV2.NewActorParser).(*actorsV2.ActorParser)
+	actor, err := p.GetActor(manifest.DatacapKey)
+	require.NoError(t, err)
+	require.NotNil(t, actor)
+
+	for _, tt := range datacapWithParamsOrReturnTests {
+		t.Run(tt.name, func(t *testing.T) {
+			rawParams, err := loadFile(manifest.DatacapKey, tt.txType, tt.key)
+			require.NoError(t, err)
+			require.NotNil(t, rawParams)
+
+			got, _, err := actor.Parse(network, height, tt.txType, &parser.LotusMessage{
+				Params: rawParams,
+			}, &parser.LotusMessageReceipt{
+				Return: nil,
+			}, cid.Undef, filTypes.EmptyTSK)
+			require.NoError(t, err)
+			require.NotNil(t, got)
+			require.Contains(t, got, tt.key, fmt.Sprintf("%s could no be found in metadata", tt.key))
+			require.NotNil(t, got[tt.key])
+		})
+	}
+}
