+++ actors/tests/paymentchannel_test.go
@@ -1,0 +1,77 @@
+package actortest
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/filecoin-project/go-state-types/manifest"
+	filTypes "github.com/filecoin-project/lotus/chain/types"
+	"github.com/ipfs/go-cid"
+	"github.com/stretchr/testify/require"
+	actorsV1 "github.com/zondax/fil-parser/actors/v1"
+	actorsV2 "github.com/zondax/fil-parser/actors/v2"
+	"github.com/zondax/fil-parser/parser"
+)
+
+var paymentChannelWithParamsOrReturnTests = []struct {
+	name   string
+	txType string
+	key    string
+}{
+	{
+		name:   "Constructor",
+		txType: parser.MethodConstructor,
+		key:    parser.ParamsKey,
+	},
+	{
+		name:   "Update Channel State",
+		txType: parser.MethodUpdateChannelState,
+		key:    parser.ParamsKey,
+	},
+}
+
+func TestActorParserV1_PaymentChannelWithParamsOrReturn(t *testing.T) {
+	p := getActorParser(actorsV1.NewActorParser).(*actorsV1.ActorParser)
+	for _, tt := range paymentChannelWithParamsOrReturnTests {
+		t.Run(tt.name, func(t *testing.T) {
+			rawParams, err := loadFile(manifest.PaychKey, tt.txType, tt.key)
+			require.NoError(t, err)
+			require.NotNil(t, rawParams)
+
+			got, err := p.ParsePaymentchannel(tt.txType, &parser.LotusMessage{
+				Params: rawParams,
+			}, &parser.LotusMessageReceipt{
+				Return: nil,
+			})
+			require.NoError(t, err)
+			require.NotNil(t, got)
+			require.Contains(t, got, tt.key, fmt.Sprintf("%s could no be found in metadata", tt.key))
+			require.NotNil(t, got[tt.key])
+		})
+	}
+}
+
+func TestActorParserV2_PaymentChannelWithParamsOrReturn(t *testing.T) {
+	p := getActorParser(actorsV2.NewActorParser).(*actorsV2.ActorParser)
+	actor, err := p.GetActor(manifest.PaychKey)
+	require.NoError(t, err)
+	require.NotNil(t, actor)
+
+	for _, tt := range paymentChannelWithParamsOrReturnTests {
+		t.Run(tt.name, func(t *testing.T) {
+			rawParams, err := loadFile(manifest.PaychKey, tt.txType, tt.key)
+			require.NoError(t, err)
+			require.NotNil(t, rawParams)
+
+			got, _, err := actor.Parse(network, height, tt.txType, &parser.LotusMessage{
+				Params: rawParams,
+			}, &parser.LotusMessageReceipt{
+				Return: nil,
+			}, cid.Undef, filTypes.EmptyTSK)
+			require.NoError(t, err)
+			require.NotNil(t, got)
+			require.Contains(t, got, tt.key, fmt.Sprintf("%s could no be found in metadata", tt.key))
+			require.NotNil(t, got[tt.key])
+		})
+	}
+}
