+++ actors/tests/testutil.go
@@ -1,4 +1,4 @@
-package actors
+package actortest
 
 import (
 	"context"
@@ -8,31 +8,26 @@ import (
 	"net/http"
 	"os"
 
+	"github.com/filecoin-project/lotus/api/client"
 	filTypes "github.com/filecoin-project/lotus/chain/types"
+	"github.com/zondax/fil-parser/actors"
 	"github.com/zondax/fil-parser/actors/cache"
 	"github.com/zondax/fil-parser/actors/cache/impl/common"
 	"github.com/zondax/fil-parser/parser"
 	helper2 "github.com/zondax/fil-parser/parser/helper"
 	"github.com/zondax/fil-parser/types"
-
-	"github.com/filecoin-project/lotus/api/client"
 	rosettaFilecoinLib "github.com/zondax/rosetta-filecoin-lib"
+	"go.uber.org/zap"
 )
 
 const (
+	network  = "mainnet"
+	dataPath = "../../data/actors"
 	testUrl  = "https://api.zondax.ch/fil/node/mainnet/rpc/v1"
-	dataPath = "../data/actors"
+	height   = int64(0)
 )
 
-func loadFile(actor, txType, name string) ([]byte, error) {
-	item, err := os.ReadFile(fmt.Sprintf("%s/%s/%s/%s", dataPath, actor, txType, name))
-	if err != nil {
-		return nil, err
-	}
-	return item, nil
-}
-
-func getActorParser() *ActorParser {
+func getActorParser(actorParserFn func(*helper2.Helper, *zap.Logger) actors.ActorParserInterface) actors.ActorParserInterface {
 	lotusClient, _, err := client.NewFullNodeRPCV1(context.Background(), testUrl, http.Header{})
 	if err != nil {
 		return nil
@@ -47,8 +42,12 @@ func getActorParser() *ActorParser {
 
 	lib := rosettaFilecoinLib.NewRosettaConstructionFilecoin(lotusClient)
 	helper := helper2.NewHelper(lib, actorsCache, lotusClient, nil)
+	logger, err := zap.NewDevelopment()
+	if err != nil {
+		return nil
+	}
 
-	return NewActorParser(helper, nil)
+	return actorParserFn(helper, logger)
 }
 
 func getParamsAndReturn(actor, txType string) ([]byte, []byte, error) {
@@ -61,6 +60,14 @@ func getParamsAndReturn(actor, txType string) ([]byte, []byte, error) {
 		return nil, nil, err
 	}
 	return rawParams, rawReturn, nil
+}
+
+func loadFile(actor, txType, name string) ([]byte, error) {
+	item, err := os.ReadFile(fmt.Sprintf("%s/%s/%s/%s", dataPath, actor, txType, name))
+	if err != nil {
+		return nil, err
+	}
+	return item, nil
 }
 
 func deserializeMessage(actor, txType string) (*parser.LotusMessage, error) {
