--- actors/v1/account_test.go
@@ -1,44 +1,0 @@
-package actors
-
-import (
-	"github.com/filecoin-project/go-state-types/manifest"
-	"github.com/stretchr/testify/require"
-	"github.com/zondax/fil-parser/parser"
-	"testing"
-)
-
-func TestActorParser_accountWithParamsAndReturn(t *testing.T) {
-	p := getActorParser()
-	tests := []struct {
-		name   string
-		txType string
-		f      func([]byte, []byte) (map[string]interface{}, error)
-	}{
-		{
-			name:   "Authenticate Message",
-			txType: parser.MethodAuthenticateMessage,
-			f:      p.authenticateMessage,
-		},
-		{
-			name:   "PubKey Address",
-			txType: parser.MethodPubkeyAddress,
-			f:      p.pubkeyAddress,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			rawParams, rawReturn, err := getParamsAndReturn(manifest.AccountKey, tt.txType)
-			require.NoError(t, err)
-			require.NotNil(t, rawParams)
-			require.NotNil(t, rawReturn)
-
-			got, err := tt.f(rawParams, rawReturn)
-			require.NoError(t, err)
-			require.NotNil(t, got)
-			require.Contains(t, got, parser.ParamsKey, "Params could no be found in metadata")
-			require.NotNil(t, got[parser.ParamsKey])
-			require.Contains(t, got, parser.ReturnKey, "Return could no be found in metadata")
-			require.NotNil(t, got[parser.ReturnKey])
-		})
-	}
-}
