--- actors/v1/cron_test.go
@@ -1,23 +1,0 @@
-package actors
-
-import (
-	"fmt"
-	"github.com/filecoin-project/go-state-types/manifest"
-	"github.com/stretchr/testify/require"
-	"github.com/zondax/fil-parser/parser"
-	"testing"
-)
-
-func TestActorParser_cronConstructor(t *testing.T) {
-	p := getActorParser()
-
-	rawParams, err := loadFile(manifest.CronKey, parser.MethodConstructor, parser.ParamsKey)
-	require.NoError(t, err)
-	require.NotNil(t, rawParams)
-
-	got, err := p.cronConstructor(rawParams)
-	require.NoError(t, err)
-	require.NotNil(t, got)
-	require.Contains(t, got, parser.ParamsKey, fmt.Sprintf("%s could no be found in metadata", parser.ParamsKey))
-	require.NotNil(t, got[parser.ParamsKey])
-}
