--- actors/v1/paymentChannel_test.go
@@ -1,45 +1,0 @@
-package actors
-
-import (
-	"fmt"
-	"github.com/filecoin-project/go-state-types/manifest"
-	"github.com/stretchr/testify/require"
-	"github.com/zondax/fil-parser/parser"
-	"testing"
-)
-
-func TestActorParser_paymentChannelWithParamsOrReturn(t *testing.T) {
-	p := getActorParser()
-	tests := []struct {
-		name   string
-		txType string
-		f      func([]byte) (map[string]interface{}, error)
-		key    string
-	}{
-		{
-			name:   "Constructor",
-			txType: parser.MethodConstructor,
-			f:      p.paymentChannelConstructor,
-			key:    parser.ParamsKey,
-		},
-		{
-			name:   "Update Channel State",
-			txType: parser.MethodUpdateChannelState,
-			f:      p.updateChannelState,
-			key:    parser.ParamsKey,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			rawParams, err := loadFile(manifest.PaychKey, tt.txType, tt.key)
-			require.NoError(t, err)
-			require.NotNil(t, rawParams)
-
-			got, err := tt.f(rawParams)
-			require.NoError(t, err)
-			require.NotNil(t, got)
-			require.Contains(t, got, tt.key, fmt.Sprintf("%s could no be found in metadata", tt.key))
-			require.NotNil(t, got[tt.key])
-		})
-	}
-}
