+++ actors/v2/multisig/values.go
@@ -29,12 +29,22 @@ import (
 	multisig15 "github.com/filecoin-project/go-state-types/builtin/v15/multisig"
 	multisig8 "github.com/filecoin-project/go-state-types/builtin/v8/multisig"
 	multisig9 "github.com/filecoin-project/go-state-types/builtin/v9/multisig"
+
+	verifreg10 "github.com/filecoin-project/go-state-types/builtin/v10/verifreg"
+	verifreg11 "github.com/filecoin-project/go-state-types/builtin/v11/verifreg"
+	verifreg12 "github.com/filecoin-project/go-state-types/builtin/v12/verifreg"
+	verifreg13 "github.com/filecoin-project/go-state-types/builtin/v13/verifreg"
+	verifreg14 "github.com/filecoin-project/go-state-types/builtin/v14/verifreg"
+	verifreg15 "github.com/filecoin-project/go-state-types/builtin/v15/verifreg"
+	verifreg8 "github.com/filecoin-project/go-state-types/builtin/v8/verifreg"
+	verifreg9 "github.com/filecoin-project/go-state-types/builtin/v9/verifreg"
+
 	"github.com/zondax/fil-parser/actors"
 	"github.com/zondax/fil-parser/parser"
 	"github.com/zondax/fil-parser/tools"
 )
 
-func (*Msig) ChangeOwnerAddressValue(network string, height int64, txMetadata string) (interface{}, error) {
+func (*Msig) ParseChangeOwnerAddressValue(network string, height int64, txMetadata string) (interface{}, error) {
 	return parseValue(txMetadata, &ChangeOwnerAddressParams{})
 }
 
@@ -102,6 +112,30 @@ func (*Msig) ParseAddSignerValue(network string, height int64, txMetadata string
 	return nil, fmt.Errorf("%w: %d", actors.ErrUnsupportedHeight, height)
 }
 
+func (*Msig) ParseAddVerifierValue(network string, height int64, txMetadata string) (interface{}, error) {
+	switch {
+	case tools.AnyIsSupported(network, height, tools.VersionsBefore(tools.V15)...):
+		return nil, fmt.Errorf("%w: %d", actors.ErrInvalidHeightForMethod, height)
+	case tools.V16.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg8.AddVerifierParams{})
+	case tools.V17.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg9.AddVerifierParams{})
+	case tools.V18.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg10.AddVerifierParams{})
+	case tools.AnyIsSupported(network, height, tools.V20, tools.V19):
+		return parseValue(txMetadata, &verifreg11.AddVerifierParams{})
+	case tools.V21.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg12.AddVerifierParams{})
+	case tools.V22.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg13.AddVerifierParams{})
+	case tools.V23.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg14.AddVerifierParams{})
+	case tools.V24.IsSupported(network, height):
+		return parseValue(txMetadata, &verifreg15.AddVerifierParams{})
+	}
+	return nil, fmt.Errorf("%w: %d", actors.ErrUnsupportedHeight, height)
+}
+
 func (*Msig) ParseApproveValue(network string, height int64, txMetadata string) (interface{}, error) {
 	var raw map[string]interface{}
 	err := json.Unmarshal([]byte(txMetadata), &raw)
@@ -154,7 +188,7 @@ func (*Msig) ParseCancelValue(network string, height int64, txMetadata string) (
 	return v, nil
 }
 
-func (*Msig) ChangeNumApprovalsThresholdValue(network string, height int64, txMetadata string) (interface{}, error) {
+func (*Msig) ParseChangeNumApprovalsThresholdValue(network string, height int64, txMetadata string) (interface{}, error) {
 	var raw map[string]interface{}
 	err := json.Unmarshal([]byte(txMetadata), &raw)
 	if err != nil {
