--- actors/v2/multisig/verifier.go
@@ -1,38 +1,0 @@
-package multisig
-
-import (
-	"fmt"
-
-	verifreg10 "github.com/filecoin-project/go-state-types/builtin/v10/verifreg"
-	verifreg11 "github.com/filecoin-project/go-state-types/builtin/v11/verifreg"
-	verifreg12 "github.com/filecoin-project/go-state-types/builtin/v12/verifreg"
-	verifreg13 "github.com/filecoin-project/go-state-types/builtin/v13/verifreg"
-	verifreg14 "github.com/filecoin-project/go-state-types/builtin/v14/verifreg"
-	verifreg15 "github.com/filecoin-project/go-state-types/builtin/v15/verifreg"
-	verifreg8 "github.com/filecoin-project/go-state-types/builtin/v8/verifreg"
-	verifreg9 "github.com/filecoin-project/go-state-types/builtin/v9/verifreg"
-	"github.com/zondax/fil-parser/actors"
-	"github.com/zondax/fil-parser/tools"
-)
-
-func (*Msig) AddVerifierValue(network string, height int64, txMetadata string) (interface{}, error) {
-	switch {
-	case tools.V16.IsSupported(network, height):
-		return parse[*verifreg8.AddVerifierParams, string](txMetadata, &verifreg8.AddVerifierParams{}, jsonUnmarshaller[*verifreg8.AddVerifierParams])
-	case tools.V17.IsSupported(network, height):
-		return parse[*verifreg9.AddVerifierParams, string](txMetadata, &verifreg9.AddVerifierParams{}, jsonUnmarshaller[*verifreg9.AddVerifierParams])
-	case tools.V18.IsSupported(network, height):
-		return parse[*verifreg10.AddVerifierParams, string](txMetadata, &verifreg10.AddVerifierParams{}, jsonUnmarshaller[*verifreg10.AddVerifierParams])
-	case tools.AnyIsSupported(network, height, tools.V20, tools.V19):
-		return parse[*verifreg11.AddVerifierParams, string](txMetadata, &verifreg11.AddVerifierParams{}, jsonUnmarshaller[*verifreg11.AddVerifierParams])
-	case tools.V21.IsSupported(network, height):
-		return parse[*verifreg12.AddVerifierParams, string](txMetadata, &verifreg12.AddVerifierParams{}, jsonUnmarshaller[*verifreg12.AddVerifierParams])
-	case tools.V22.IsSupported(network, height):
-		return parse[*verifreg13.AddVerifierParams, string](txMetadata, &verifreg13.AddVerifierParams{}, jsonUnmarshaller[*verifreg13.AddVerifierParams])
-	case tools.V23.IsSupported(network, height):
-		return parse[*verifreg14.AddVerifierParams, string](txMetadata, &verifreg14.AddVerifierParams{}, jsonUnmarshaller[*verifreg14.AddVerifierParams])
-	case tools.V24.IsSupported(network, height):
-		return parse[*verifreg15.AddVerifierParams, string](txMetadata, &verifreg15.AddVerifierParams{}, jsonUnmarshaller[*verifreg15.AddVerifierParams])
-	}
-	return nil, fmt.Errorf("%w: %d", actors.ErrInvalidHeightForMethod, height)
-}
